requirements.txt
yfinance==0.2.36
newsapi-python==0.2.7
vaderSentiment==3.3.2
pandas==2.1.0
numpy==1.26.0
matplotlib==3.8.0
plotly==5.17.0
streamlit==1.28.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn==1.3.0
